+++ src/dedalus_labs/__init__.py
@@ -21,6 +21,7 @@ from ._models import BaseModel
 from ._version import __title__, __version__
 from ._response import APIResponse as APIResponse, AsyncAPIResponse as AsyncAPIResponse
 from ._constants import DEFAULT_TIMEOUT, DEFAULT_MAX_RETRIES, DEFAULT_CONNECTION_LIMITS
+from .lib.runner import DedalusRunner
 from ._exceptions import (
     APIError,
     DedalusError,
@@ -34,13 +35,13 @@ from ._exceptions import (
     AuthenticationError,
     InternalServerError,
     PermissionDeniedError,
+    SchemaProcessingError,
     UnprocessableEntityError,
     APIResponseValidationError,
-    SchemaProcessingError,
 )
 from ._base_client import DefaultHttpxClient, DefaultAioHttpClient, DefaultAsyncHttpxClient
 from ._utils._logs import setup_logging as _setup_logging
-from .lib.runner import DedalusRunner
+from .types.dedalus_model import DedalusModel
 
 __all__ = [
     "types",
@@ -84,6 +85,7 @@ __all__ = [
     "DefaultAsyncHttpxClient",
     "DefaultAioHttpClient",
     "DedalusRunner",
+    "DedalusModel",
     "SchemaProcessingError",
 ]